air_quality_backend/predictions/serializers.py
from rest_framework import serializers
from .models import PredictionResult, DataPoint

class PredictionResultSerializer(serializers.ModelSerializer):
    class Meta:
        model = PredictionResult
        fields = '__all__'

class DataPointSerializer(serializers.ModelSerializer):
    class Meta:
        model = DataPoint
        fields = '__all__'
