requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
h5py==3.9.0
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
python-dateutil==2.8.2
pytz==2023.3
